🚚 Move version info modal out of its folder
