+++ src/modals/index.js
@@ -1,3 +1,3 @@
 export {default as NewVersionModal} from './NewVersionModal/NewVersionFlow';
-export {default as VersionInfoModal} from './VersionInfoModal/VersionInfoModal';
+export {default as VersionInfoModal} from './VersionInfoModal';
 export {default as EditDocumentModal} from './EditDocumentModal';
